requirements.txt
llama-cpp-python
rich
typer
